+++ docs/changelog.rst
@@ -3,12 +3,16 @@
 
 ### v1.0.2 (unreleased)
 
-- *2020-09-17* **ScanPipe** -- Always return the Pipeline subclass/implementation
+- **ScanPipe** -- Add a createproject management command.
+  Fix for https://github.com/nexB/scancode.io/issues/13
+
+- **ScanPipe** -- Always return the Pipeline subclass/implementation
   from the module inspection. Fix for https://github.com/nexB/scancode.io/issues/11
+
 
 ### v1.0.1 (2020-09-12)
 
-- *2020-09-11* **ScanPipe** -- Do not fail when collecting system packages in
+- **ScanPipe** -- Do not fail when collecting system packages in
   Ubuntu docker images for layers that do not install packages by updating to a
   newer version of ScanCode Toolkit. Fix for https://github.com/nexB/scancode.io/issues/1